+++ packages/editor/js/extensions/libs/block-card/inner-blocks/hooks/use-available-items.js
@@ -34,10 +34,10 @@ export const useAvailableItems = ({
 	const { getBlockType } = select('core/blocks');
 	const { getAllowedBlocks, getSelectedBlock } = select('core/block-editor');
 	const allowedBlockTypes = getAllowedBlocks(clientId);
-	const innerBlocks = getSelectedBlock().innerBlocks;
+	const { innerBlocks, attributes } = getSelectedBlock();
 
 	return useMemo(() => {
-		let forces: Array<InnerBlockModel> = [];
+		const forces: Array<InnerBlockModel> = [];
 		const blocks: Array<InnerBlockModel> = [];
 		const elements: Array<InnerBlockModel> = [];
 
@@ -88,7 +88,11 @@ export const useAvailableItems = ({
 		// Appending allowed block types of WordPress selected block ...
 		appendBlocks(allowedBlockTypes);
 
-		forces = modifyItemsPriority(innerBlocks, forces);
+		const { forcesItems, customizedInnerBlocks } = modifyItemsPriority(
+			innerBlocks,
+			forces,
+			attributes
+		);
 
 		// Appending forces into repeater state.
 		if (forces.length) {
@@ -97,7 +101,12 @@ export const useAvailableItems = ({
 
 			setBlockClientInners({
 				clientId,
-				inners: calculateInners(forces, inners, maxItems),
+				inners: calculateInners({
+					inners,
+					maxItems,
+					forcesItems,
+					customizedInnerBlocks,
+				}),
 			});
 		}
 
@@ -108,15 +117,66 @@ export const useAvailableItems = ({
 
 export const modifyItemsPriority = (
 	appendedItems: Array<InnerBlockModel>,
-	forcesItems: Array<InnerBlockModel>
-): Array<InnerBlockModel> => {
+	forcesItems: Array<InnerBlockModel>,
+	attributes: Object
+): {
+	forcesItems: Array<InnerBlockModel>,
+	customizedInnerBlocks: Array<InnerBlockModel>,
+} => {
+	const customizedInnerBlocks: Array<InnerBlockModel> = [];
+	const processedItems: Set<string> = new Set();
+
 	const forcesItemsProcessor = (blockType: InnerBlockModel) => {
 		for (let index = 0; index < forcesItems.length; index++) {
 			const elementType = forcesItems[index];
+			const priority = elementType.settings?.priority || 10;
+
+			if (
+				Object.keys(
+					(
+						attributes?.blockeraInnerBlocks?.value[
+							elementType.name
+						] || {}
+					)?.attributes || {}
+				).length
+			) {
+				let newPriority = priority - 1 === -1 ? 0 : priority - 1;
+
+				if (elementType.name === blockType.name && newPriority > 0) {
+					newPriority -= 1;
+				}
+
+				const foundedItemIndex = customizedInnerBlocks.findIndex(
+					(block: InnerBlockModel) => block.name === blockType?.name
+				);
+				const hasElement = foundedItemIndex !== -1;
+
+				if (hasElement) {
+					const foundedItem = customizedInnerBlocks[foundedItemIndex];
+
+					if (foundedItem.settings.priority - 1 === -1) {
+						foundedItem.settings.priority = 0;
+					} else {
+						foundedItem.settings.priority -= 1;
+					}
+
+					customizedInnerBlocks[foundedItemIndex] = foundedItem;
+				} else {
+					if (!processedItems.has(elementType.name)) {
+						customizedInnerBlocks.push({
+							...elementType,
+							settings: {
+								...elementType.settings,
+								priority: newPriority,
+							},
+						});
+					}
+
+					processedItems.add(elementType.name);
+				}
+			}
 
 			if (elementType.name === blockType.name) {
-				const priority = elementType.settings?.priority || 10;
-
 				forcesItems[index] = {
 					...elementType,
 					settings: {
@@ -141,17 +201,28 @@ export const modifyItemsPriority = (
 		}
 	}
 
-	return forcesItems;
+	return {
+		forcesItems,
+		customizedInnerBlocks,
+	};
 };
 
-const calculateInners = (
-	forces: Array<InnerBlockModel>,
+const calculateInners = ({
+	inners,
+	maxItems,
+	forcesItems,
+	customizedInnerBlocks,
+}: {
 	inners: InnerBlocks,
-	maxItems?: number | void
-) => {
-	forces = getSortedObject(
-		mergeObject(
-			forces.reduce(
+	maxItems?: number | void,
+	forcesItems: Array<InnerBlockModel>,
+	customizedInnerBlocks: Array<InnerBlockModel>,
+}) => {
+	const normalizedInners = (
+		items: Array<InnerBlockModel>
+	): { [key: string]: InnerBlockModel } => {
+		return getSortedObject(
+			items.reduce(
 				(accumulator, innerBlock) => ({
 					...accumulator,
 					[innerBlock?.name]: {
@@ -162,27 +233,78 @@ const calculateInners = (
 				}),
 				{}
 			),
-			inners
-		),
-		'settings',
-		10
-	);
+			'settings',
+			10
+		);
+	};
+	const limitedByMaxItems = (
+		items: {
+			[key: string]: InnerBlockModel,
+		},
+		maxItems: number,
+		registeredItems?: Array<InnerBlockModel>
+	): { [key: string]: InnerBlockModel } => {
+		const limitedForces: { [key: string]: InnerBlockModel } = {};
 
-	if (!maxItems) {
-		return forces;
-	}
+		for (const key in items) {
+			const force = items[key];
 
-	const limitedForces: { [key: string]: InnerBlockModel } = {};
+			if (
+				registeredItems &&
+				registeredItems?.find((item) => item.name === force.name)
+			) {
+				continue;
+			}
 
-	for (const key in forces) {
-		const force = forces[key];
+			if (Object.keys(limitedForces).length >= maxItems) {
+				break;
+			}
 
-		if (Object.keys(limitedForces).length >= maxItems) {
-			break;
+			limitedForces[force.name] = force;
 		}
 
-		limitedForces[force.name] = force;
+		return limitedForces;
+	};
+
+	if (maxItems) {
+		if (customizedInnerBlocks.length === maxItems) {
+			return mergeObject(normalizedInners(customizedInnerBlocks), inners);
+		}
+
+		if (customizedInnerBlocks.length < maxItems) {
+			const limitedForcesInnerBlocks = limitedByMaxItems(
+				normalizedInners(forcesItems),
+				maxItems - customizedInnerBlocks.length,
+				customizedInnerBlocks
+			);
+
+			return mergeObject(
+				{
+					...normalizedInners(customizedInnerBlocks),
+					...limitedForcesInnerBlocks,
+				},
+				inners
+			);
+		}
+
+		if (!customizedInnerBlocks.length) {
+			return mergeObject(
+				limitedByMaxItems(normalizedInners(forcesItems), maxItems),
+				inners
+			);
+		}
+
+		return mergeObject(normalizedInners(customizedInnerBlocks), inners);
 	}
 
-	return limitedForces;
+	if (customizedInnerBlocks.length) {
+		const mergedItems = mergeObject(
+			normalizedInners(customizedInnerBlocks),
+			normalizedInners(forcesItems)
+		);
+
+		return mergeObject(mergedItems, inners);
+	}
+
+	return mergeObject(normalizedInners(forcesItems), inners);
 };